debbugers: Fix void realgud:run-process error in gdb and gub.
Fixes <https://github.com/realgud/realgud/issues/269>.

The 'realgud:run-process' function was not defined in the context of
the gdb and gub debuggers, causing an error when running 'M-x
realgud:gdb', for example.

These are the important bits that fix the above problem:

* realgud/debugger/gdb/gdb.el: Require "../../common/run".
* realgud/debugger/gub/gub.el: Likewise.

The following changes are cosmetic, non-functional:

* realgud/debugger/gdb/gdb.el (realgud:run-process): Fix "file"
reference.
* realgud/debugger/gub/gub.el (realgud:run-process): Likewise.
* realgud/debugger/pdb/pdb.el (realgud:run-process): Likewise.
* realgud/debugger/trepan2/trepan2.el (realgud:run-process): Likewise.
* realgud/debugger/trepan3k/trepan3k.el (realgud:run-process): Likewise.